+++ rustdf/src/sim/dia.rs
@@ -22,9 +22,9 @@ pub struct TimsTofSyntheticsFrameBuilderDIA {
     pub transmission_settings: TimsTransmissionDIA,
     pub fragmentation_settings: TimsTofCollisionEnergyDIA,
     pub fragment_ions:
-        Option<BTreeMap<(u32, i8, i8), (PeptideProductIonSeriesCollection, Vec<MzSpectrum>)>>,
+        Option<BTreeMap<(u32, i8, i32), (PeptideProductIonSeriesCollection, Vec<MzSpectrum>)>>,
     pub fragment_ions_annotated: Option<
-        BTreeMap<(u32, i8, i8), (PeptideProductIonSeriesCollection, Vec<MzSpectrumAnnotated>)>,
+        BTreeMap<(u32, i8, i32), (PeptideProductIonSeriesCollection, Vec<MzSpectrumAnnotated>)>,
     >,
 }
 
@@ -435,7 +435,7 @@ impl TimsTofSyntheticsFrameBuilderDIA {
         &self,
         frame_id: u32,
         fragment_ions: &BTreeMap<
-            (u32, i8, i8),
+            (u32, i8, i32),
             (PeptideProductIonSeriesCollection, Vec<MzSpectrum>),
         >,
         mz_noise_fragment: bool,
@@ -540,7 +540,7 @@ impl TimsTofSyntheticsFrameBuilderDIA {
                     let collision_energy = self
                         .fragmentation_settings
                         .get_collision_energy(frame_id as i32, *scan as i32);
-                    let collision_energy_quantized = (collision_energy * 1e3).round() as i8;
+                    let collision_energy_quantized = (collision_energy * 1e3).round() as i32;
 
                     // get charge state for the ion
                     let charge_state = charges.get(index).unwrap();
@@ -630,7 +630,7 @@ impl TimsTofSyntheticsFrameBuilderDIA {
         &self,
         frame_id: u32,
         fragment_ions: &BTreeMap<
-            (u32, i8, i8),
+            (u32, i8, i32),
             (PeptideProductIonSeriesCollection, Vec<MzSpectrumAnnotated>),
         >,
         mz_noise_fragment: bool,
@@ -736,7 +736,7 @@ impl TimsTofSyntheticsFrameBuilderDIA {
                     let collision_energy = self
                         .fragmentation_settings
                         .get_collision_energy(frame_id as i32, *scan as i32);
-                    let collision_energy_quantized = (collision_energy * 1e3).round() as i8;
+                    let collision_energy_quantized = (collision_energy * 1e3).round() as i32;
 
                     let charge_state = charges.get(index).unwrap();
                     let maybe_value = fragment_ions.get(&(
